Changed to new Hooks method instead of export.lua
Updated Theme icons - thanks to Kobalt
Updated DCS integration - thanks to Dexter and ED
